refactor(rpc): extract Stellar mock fixtures into mock_stellar module
Move deterministic Horizon/RPC mock generators and ledger constants out of
StellarRpcClient so the client focuses on transport while fixtures live in a
dedicated module. StellarRpcClient delegates mock branches to super::mock_stellar.

This is a stepping stone toward a full Real/Mock trait-backed client; the
public API of StellarRpcClient is unchanged.

Made-with: Cursor
